+++ doc/bin/docker-compose-down.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-cd doc/docker || exit 1
-
-docker compose -f compose-local.yaml down -v