+++ doc/bin/docker-compose-local-rebuild.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-set -euo pipefail
-
-./mvnw clean install -Pdocker
-doc/bin/docker-compose-local-restart.sh