+++ doc/bin/test-with-compose-local-rebuild.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-# stop on every error
-set -e
-
-doc/bin/docker-compose-local-rebuild.sh
-doc/bin/test-with-compose-local.sh